+++ lib/src/ui/dynamic_height_container.dart
@@ -2,23 +2,27 @@ import 'package:day_schedule_list/src/ui/interval_containers/appointment_contain
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
-typedef CanUpdateToHeight = bool Function(double);
-typedef UpdateCallback = void Function();
-typedef UpdateHeightCallback = void Function(double height);
+typedef CanUpdateToHeight = bool Function(double, HeightUpdateFrom);
+typedef UpdateCallback = void Function(HeightUpdateFrom from);
+typedef UpdateHeightCallback = void Function(double height, HeightUpdateFrom from);
+
+enum HeightUpdateFrom {
+  top, bottom
+}
 
 class DynamicHeightContainer extends StatefulWidget {
   const DynamicHeightContainer(
       {required this.canUpdateHeightTo,
+      required this.onUpdateStart,
+      required this.onNewUpdate,
+      required this.onUpdateEnd,
+      required this.onUpdateCancel,
       required this.currentHeight,
       required this.child,
       this.dragIndicatorColor,
       this.dragIndicatorBorderColor,
       this.dragIndicatorBorderWidth,
       this.updateStep,
-      this.onUpdateStart,
-      this.onNewUpdate,
-      this.onUpdateEnd,
-      this.onUpdateCancel,
       Key? key})
       : assert(
           dragIndicatorBorderWidth == null || dragIndicatorBorderWidth > 0,
@@ -49,16 +53,16 @@ class DynamicHeightContainer extends StatefulWidget {
   final double currentHeight;
 
   ///Callback called when height update action starts.
-  final UpdateCallback? onUpdateStart;
+  final UpdateCallback onUpdateStart;
 
   ///Callback called everytime height value changes.
-  final UpdateHeightCallback? onNewUpdate;
+  final UpdateHeightCallback onNewUpdate;
 
   ///Callback called when height update action ends
-  final UpdateHeightCallback? onUpdateEnd;
+  final UpdateHeightCallback onUpdateEnd;
 
   ///Callback called when height update action ends
-  final UpdateCallback? onUpdateCancel;
+  final UpdateCallback onUpdateCancel;
 
   ///Function to determine if [child] height should or not change to a new one.
   final CanUpdateToHeight canUpdateHeightTo;
@@ -69,73 +73,74 @@ class DynamicHeightContainer extends StatefulWidget {
 
 class _DynamicHeightContainerState extends State<DynamicHeightContainer> {
   ///Current widget height
-  late ValueNotifier<double> _currentHeight;
+  late double _currentHeight;
 
   ///The sum of last vertical drag gesture [DragUpdateDetails.delta.dy]  that does not
   ///caused View height to change because it is less than [widget.updateStep]
   double _pendingDeltaYForUpdateStep = 0;
   Offset _oldOffsetFromOrigin = Offset.zero;
   bool _didMove = false;
-
+  HeightUpdateFrom?  _updateFrom;
   @override
   void initState() {
-    _currentHeight = ValueNotifier(widget.currentHeight);
+    _currentHeight = widget.currentHeight;
     super.initState();
-  }
-
-  @override
-  void dispose() {
-    _currentHeight.dispose();
-    super.dispose();
   }
 
   @override
   void didUpdateWidget(covariant DynamicHeightContainer oldWidget) {
     super.didUpdateWidget(oldWidget);
-    _currentHeight.value = widget.currentHeight;
+    _currentHeight = widget.currentHeight;
+    _updateFrom = null;
   }
 
   @override
   Widget build(BuildContext context) {
-    return DragIndicatorWidget.bottom(
-      onLongPressDown: onLongPressDown,
-      onLongPressStart: onLongPressStart,
-      onLongPressEnd: onLongPressEnd,
-      onLongPressMoveUpdate: onLongPressMoveUpdate,
-      dragIndicatorColor: widget.dragIndicatorColor,
-      dragIndicatorBorderColor: widget.dragIndicatorBorderColor,
-      dragIndicatorBorderWidth: widget.dragIndicatorBorderWidth,
-      child: ValueListenableBuilder<double>(
-        valueListenable: _currentHeight,
-        builder: (context, value, child) {
-          return SizedBox(
-            height: value,
-            child: child,
-          );
-        },
-        child: widget.child,
-      ),
+    return Stack(
+      clipBehavior: Clip.none,
+      children: [
+        SizedBox(
+          height: _currentHeight,
+          child: widget.child,
+        ),
+        DragIndicatorWidget.top(
+          onLongPressDown: () => onLongPressDown(HeightUpdateFrom.top),
+          onLongPressStart: (_){},//onLongPressStart,
+          onLongPressEnd: onLongPressEnd,
+          onLongPressMoveUpdate: onLongPressMoveUpdate,
+          dragIndicatorColor: widget.dragIndicatorColor,
+          dragIndicatorBorderColor: widget.dragIndicatorBorderColor,
+          dragIndicatorBorderWidth: widget.dragIndicatorBorderWidth,
+        ),
+        DragIndicatorWidget.bottom(
+          onLongPressDown: () => onLongPressDown(HeightUpdateFrom.bottom),
+          onLongPressStart: (_){},//onLongPressStart,
+          onLongPressEnd: onLongPressEnd,
+          onLongPressMoveUpdate: onLongPressMoveUpdate,
+          dragIndicatorColor: widget.dragIndicatorColor,
+          dragIndicatorBorderColor: widget.dragIndicatorBorderColor,
+          dragIndicatorBorderWidth: widget.dragIndicatorBorderWidth,
+        )
+      ],
     );
   }
 
-  void onLongPressDown() {
+  void onLongPressDown(HeightUpdateFrom from) {
     HapticFeedback.heavyImpact();
     _pendingDeltaYForUpdateStep = 0;
     _oldOffsetFromOrigin = Offset.zero;
     _didMove = false;
-  }
-
-  void onLongPressStart(LongPressStartDetails details) {
+    _updateFrom = from;
     _informUpdateStart();
   }
 
   void onLongPressMoveUpdate(LongPressMoveUpdateDetails details) {
-    _didMove = true;
     double? updateStep = widget.updateStep;
     final offsetY = details.offsetFromOrigin.dy - _oldOffsetFromOrigin.dy;
     if (updateStep != null) {
       final double nextPendingIncrement = _pendingDeltaYForUpdateStep + offsetY;
       if (nextPendingIncrement.abs() >= updateStep) {
+        _didMove = true;
         _performIncrementBy(
           updateStep * (nextPendingIncrement / nextPendingIncrement.abs()),
         );
@@ -144,6 +149,7 @@ class _DynamicHeightContainerState extends State<DynamicHeightContainer> {
         _pendingDeltaYForUpdateStep = nextPendingIncrement;
       }
     } else {
+      _didMove = true;
       _performIncrementBy(offsetY);
     }
     _oldOffsetFromOrigin = details.offsetFromOrigin;
@@ -164,35 +170,42 @@ class _DynamicHeightContainerState extends State<DynamicHeightContainer> {
   }
 
   void _informUpdateStart() {
-    if (widget.onUpdateStart != null) {
-      widget.onUpdateStart!();
+    final from = _updateFrom;
+    if (from != null) {
+      widget.onUpdateStart(from);
     }
   }
 
   void _informUpdateEnd() {
-    if (widget.onUpdateEnd != null) {
-      widget.onUpdateEnd!(_currentHeight.value);
+    final from = _updateFrom;
+    if (from != null) {
+      widget.onUpdateEnd(_currentHeight, from);
     }
   }
 
   void _informNewUpdate() {
-    if (widget.onNewUpdate != null) {
-      widget.onNewUpdate!(_currentHeight.value);
+    final from = _updateFrom;
+    if (from != null) {
+      widget.onNewUpdate(_currentHeight, from);
     }
   }
 
   void _informUpdateCancel() {
-    if (widget.onUpdateCancel != null) {
-      widget.onUpdateCancel!();
+    final from = _updateFrom;
+    if (from != null) {
+      widget.onUpdateCancel(from);
     }
   }
 
   void _performIncrementBy(double value) {
-    final localCurrentHeight = _currentHeight.value;
-    final finalHeight = localCurrentHeight + value;
-    if (widget.canUpdateHeightTo(finalHeight)) {
-      _currentHeight.value = finalHeight;
-      _informNewUpdate();
+    final from = _updateFrom;
+    if(from != null) {
+      final localCurrentHeight = _currentHeight;
+      final finalHeight = localCurrentHeight + value*(from == HeightUpdateFrom.bottom ? 1 : -1);
+      if (widget.canUpdateHeightTo(finalHeight, from)) {
+        _currentHeight = finalHeight;
+        _informNewUpdate();
+      }
     }
   }
 }