+++ lib/src/ui/interval_containers/appointment_container/appointment_container.dart
@@ -4,21 +4,20 @@ import '../../day_schedule_list_widget.dart';
 import '../../dynamic_height_container.dart';
 import '../../../models/schedule_item_position.dart';
 import '../dynamic_position_container.dart';
-import 'appointment_time_of_day_indicator_widget.dart';
+
+typedef AppointmentUpdatePositionStartCallback = void Function(AppointmentUpdatingMode mode);
 
 class AppointmentContainer extends StatefulWidget {
   const AppointmentContainer({
     required this.position,
     required this.timeIndicatorsInset,
-    required this.updateHeightStep,
+    required this.updateStep,
     required this.canUpdateHeightTo,
-    required this.onUpdateHeightEnd,
     required this.canUpdatePositionTo,
     required this.onUpdatePositionEnd,
     required this.onNewUpdatePosition,
     required this.onUpdatePositionStart,
     required this.onUpdatePositionCancel,
-    required this.endTimeOfDayForPossibleNewHeight,
     required this.child,
     this.dragIndicatorBorderWidth,
     this.dragIndicatorColor,
@@ -28,17 +27,15 @@ class AppointmentContainer extends StatefulWidget {
 
   final ScheduleItemPosition position;
   final Widget child;
-  final double updateHeightStep;
+  final double updateStep;
   final double timeIndicatorsInset;
 
   final CanUpdateToHeight canUpdateHeightTo;
-  final UpdateHeightCallback onUpdateHeightEnd;
-  final TimeOfDay Function(double newHeight) endTimeOfDayForPossibleNewHeight;
 
   final CanUpdateToPosition canUpdatePositionTo;
   final UpdatePositionCallback onUpdatePositionEnd;
   final UpdatePositionCallback onNewUpdatePosition;
-  final void Function() onUpdatePositionStart;
+  final AppointmentUpdatePositionStartCallback onUpdatePositionStart;
   final void Function() onUpdatePositionCancel;
 
   final Color? dragIndicatorColor;
@@ -51,20 +48,15 @@ class AppointmentContainer extends StatefulWidget {
 
 class _AppointmentContainerState extends State<AppointmentContainer> {
   late ValueNotifier<AppointmentUpdatingMode> _updateMode;
-  late ValueNotifier<TimeOfDay> _possibleNewEndDate;
-
   @override
   void initState() {
     _updateMode = ValueNotifier(AppointmentUpdatingMode.none);
-    _possibleNewEndDate = ValueNotifier(
-        widget.endTimeOfDayForPossibleNewHeight(widget.position.height));
     super.initState();
   }
 
   @override
   void dispose() {
     _updateMode.dispose();
-    _possibleNewEndDate.dispose();
     super.dispose();
   }
 
@@ -74,133 +66,51 @@ class _AppointmentContainerState extends State<AppointmentContainer> {
       top: widget.position.top,
       right: 0,
       left: 0,
-      child: Stack(
-        clipBehavior: Clip.none,
-        children: [
-          ValueListenableBuilder(
-            valueListenable: _updateMode,
-            builder: (context, value, child) {
-              return Visibility(
-                visible: value == AppointmentUpdatingMode.changeHeight,
-                child: child!,
-              );
-            },
-            child: ValueListenableBuilder<TimeOfDay>(
-              valueListenable: _possibleNewEndDate,
-              builder: (context, endDate, child) {
-                return AppointmentTimeOfDayIndicatorWidget.end(
-                  time: endDate,
-                  timeIndicatorsInset: widget.timeIndicatorsInset,
+      child: Padding(
+        padding: const EdgeInsets.only(
+          left: DayScheduleListWidget.intervalContainerLeftInset,
+        ),
+        child: DynamicTopPositionContainer(
+          position: widget.position,
+          canUpdatePositionTo: widget.canUpdatePositionTo,
+          onNewPositionUpdate: _onNewPositionUpdate,
+          onUpdatePositionEnd: _onPositionUpdateEnd,
+          onUpdatePositionCancel: _onUpdatePositionCancel,
+          onUpdatePositionStart: _onUpdatePositionStart,
+          child: DynamicHeightContainer(
+            currentHeight: widget.position.height,
+            updateStep: widget.updateStep,
+            canUpdateHeightTo: widget.canUpdateHeightTo,
+            dragIndicatorBorderColor: widget.dragIndicatorBorderColor,
+            dragIndicatorBorderWidth: widget.dragIndicatorBorderWidth,
+            dragIndicatorColor: widget.dragIndicatorColor,
+            onUpdateEnd: _onUpdateHeightEnd,
+            onUpdateStart: _onUpdateHeightStart,
+            onUpdateCancel: _onUpdateHeightCancel,
+            onNewUpdate: _onNewUpdateHeight,
+            child: ValueListenableBuilder<AppointmentUpdatingMode>(
+              valueListenable: _updateMode,
+              builder: (context, value, child) {
+                return Opacity(
+                  opacity:
+                  value != AppointmentUpdatingMode.none ? 0.5 : 1,
+                  child: child!,
                 );
               },
-            ),
-          ),
-          Padding(
-            padding: const EdgeInsets.only(
-              left: DayScheduleListWidget.intervalContainerLeftInset,
-            ),
-            child: DynamicTopPositionContainer(
-              position: widget.position,
-              canUpdatePositionTo: widget.canUpdatePositionTo,
-              onNewPositionUpdate: _onNewPositionUpdate,
-              onUpdatePositionEnd: _onPositionUpdateEnd,
-              onUpdatePositionCancel: _onUpdateTopCancel,
-              onUpdatePositionStart: _onUpdateTopStart,
-              child: DynamicHeightContainer(
-                currentHeight: widget.position.height,
-                updateStep: widget.updateHeightStep,
-                canUpdateHeightTo: widget.canUpdateHeightTo,
-                dragIndicatorBorderColor: widget.dragIndicatorBorderColor,
-                dragIndicatorBorderWidth: widget.dragIndicatorBorderWidth,
-                dragIndicatorColor: widget.dragIndicatorColor,
-                onUpdateEnd: _onUpdateHeightEnd,
-                onUpdateStart: _onUpdateHeightStart,
-                onUpdateCancel: _onUpdateHeightCancel,
-                onNewUpdate: _onNewUpdateHeight,
-                child: ValueListenableBuilder<AppointmentUpdatingMode>(
-                  valueListenable: _updateMode,
-                  builder: (context, value, child) {
-                    return Opacity(
-                      opacity:
-                          value == AppointmentUpdatingMode.changeTop ? 0.5 : 1,
-                      child: child!,
-                    );
-                  },
-                  child: FractionallySizedBox(
-                    widthFactor: 1,
-                    child: widget.child,
-                  ),
-                ),
+              child: FractionallySizedBox(
+                widthFactor: 1,
+                child: widget.child,
               ),
             ),
-          )
-        ],
+          ),
+        ),
       ),
     );
-    // return ValueListenableBuilder<ScheduleItemPosition>(
-    //   valueListenable: _currentPosition,
-    //   builder: (
-    //     BuildContext context,
-    //     ScheduleItemPosition position,
-    //     Widget? child,
-    //   ) {
-    //     return Positioned(
-    //       top: position.top,
-    //       right: 0,
-    //       left: 0,
-    //       child: child!,
-    //     );
-    //   },
-    //   child: Stack(
-    //     clipBehavior: Clip.none,
-    //     children: [
-    //       AppointmentTimeOfDayIndicatorWidget.start(
-    //         timeIndicatorsInset: widget.timeIndicatorsInset,
-    //         updateMode: _updateMode,
-    //       ),
-    //       AppointmentTimeOfDayIndicatorWidget.end(
-    //         timeIndicatorsInset: widget.timeIndicatorsInset,
-    //         updateMode: _updateMode,
-    //       ),
-    //       Padding(
-    //         padding: const EdgeInsets.only(
-    //           left: DayScheduleList.intervalContainerLeftInset,
-    //         ),
-    //         child: DynamicTopPositionContainer(
-    //           top: widget.position.top,
-    //           canUpdateTopTo: widget.canUpdateTopTo,
-    //           onNewUpdate: _onNewTopUpdate,
-    //           onUpdateEnd: _onTopUpdateEnd,
-    //           onUpdateStart: _onUpdateTopStart,
-    //           onUpdateCancel: _onUpdateTopCancel,
-    //           child: DynamicHeightContainer(
-    //             currentHeight: widget.position.height,
-    //             updateStep: widget.updateHeightStep,
-    //             canUpdateHeightTo: widget.canUpdateHeightTo,
-    //             onUpdateEnd: _onUpdateHeightEnd,
-    //             onUpdateStart: _onUpdateHeightStart,
-    //             onUpdateCancel: _onUpdateHeightCancel,
-    //             child: ValueListenableBuilder<AppointmentUpdatingMode>(
-    //               valueListenable: _updateMode,
-    //               builder: (context, value, child) {
-    //                 return FractionallySizedBox(
-    //                   widthFactor:
-    //                       value == AppointmentUpdatingMode.changeTop ? 1.02 : 1,
-    //                   child: widget.child,
-    //                 );
-    //               },
-    //             ),
-    //           ),
-    //         ),
-    //       )
-    //     ],
-    //   ),
-    // );
   }
 
-  void _onUpdateTopStart() {
-    _updateModeToChangingTop();
-    widget.onUpdatePositionStart();
+  void _onUpdatePositionStart() {
+    _updateModeTo(AppointmentUpdatingMode.changePosition);
+    widget.onUpdatePositionStart(_updateMode.value);
   }
 
   void _onNewPositionUpdate(ScheduleItemPosition newPosition) {
@@ -208,58 +118,70 @@ class _AppointmentContainerState extends State<AppointmentContainer> {
   }
 
   void _onPositionUpdateEnd(ScheduleItemPosition newPosition) {
-    _updateModeTofinishChangeTop();
+    _updateModeToNoneFrom(AppointmentUpdatingMode.changePosition);
     widget.onUpdatePositionEnd(newPosition);
   }
 
-  void _onUpdateTopCancel() {
-    _updateModeTofinishChangeTop();
+  void _onUpdatePositionCancel() {
+    _updateModeToNoneFrom(AppointmentUpdatingMode.changePosition);
     widget.onUpdatePositionCancel();
   }
 
-  void _onUpdateHeightStart() {
-    _possibleNewEndDate.value =
-        widget.endTimeOfDayForPossibleNewHeight(widget.position.height);
-    _updateModeToChangingHeight();
+  void _onUpdateHeightStart(HeightUpdateFrom from) {
+    _updateModeTo(AppointmentUpdatingModeFromHeightUpdate.create(from));
+    widget.onUpdatePositionStart(_updateMode.value);
   }
 
-  void _onUpdateHeightEnd(double newHeight) async {
-    _updateModeTofinishChangeHeight();
-    widget.onUpdateHeightEnd(newHeight);
+  void _onNewUpdateHeight(double newHeight, HeightUpdateFrom from) {
+    widget.onNewUpdatePosition(_newPositionFor(newHeight, from));
   }
 
-  void _onNewUpdateHeight(double newHeight) {
-    _possibleNewEndDate.value =
-        widget.endTimeOfDayForPossibleNewHeight(newHeight);
+  void _onUpdateHeightEnd(double newHeight, HeightUpdateFrom from) async {
+    _updateModeToNoneFrom(AppointmentUpdatingModeFromHeightUpdate.create(from));
+    widget.onUpdatePositionEnd(_newPositionFor(newHeight, from));
   }
 
-  void _onUpdateHeightCancel() {
-    _updateModeTofinishChangeHeight();
+  void _onUpdateHeightCancel(HeightUpdateFrom from) {
+    _updateModeToNoneFrom(AppointmentUpdatingModeFromHeightUpdate.create(from));
+    widget.onUpdatePositionCancel();
   }
 
-  void _updateModeToChangingTop() {
+  void _updateModeTo(AppointmentUpdatingMode mode) {
     if (_updateMode.value == AppointmentUpdatingMode.none) {
-      _updateMode.value = AppointmentUpdatingMode.changeTop;
+      _updateMode.value = mode;
     }
   }
 
-  void _updateModeToChangingHeight() {
-    if (_updateMode.value == AppointmentUpdatingMode.none) {
-      _updateMode.value = AppointmentUpdatingMode.changeHeight;
-    }
-  }
-
-  void _updateModeTofinishChangeTop() {
-    if (_updateMode.value == AppointmentUpdatingMode.changeTop) {
+  void _updateModeToNoneFrom(AppointmentUpdatingMode mode) {
+    if (_updateMode.value == mode) {
       _updateMode.value = AppointmentUpdatingMode.none;
     }
   }
 
-  void _updateModeTofinishChangeHeight() {
-    if (_updateMode.value == AppointmentUpdatingMode.changeHeight) {
-      _updateMode.value = AppointmentUpdatingMode.none;
+  ScheduleItemPosition _newPositionFor(double newHeight, HeightUpdateFrom from) {
+    switch(from){
+      case HeightUpdateFrom.top:
+        final deltaHeight = newHeight - widget.position.height;
+        final newTop = widget.position.top - deltaHeight;
+        return ScheduleItemPosition(
+          top: newTop,
+          height: newHeight,
+        );
+      case HeightUpdateFrom.bottom:
+        return widget.position.withNewHeight(newHeight);
     }
   }
 }
 
-enum AppointmentUpdatingMode { changeTop, changeHeight, none }
+enum AppointmentUpdatingMode { changePosition, changeHeight, changeTop, none }
+
+extension AppointmentUpdatingModeFromHeightUpdate on AppointmentUpdatingMode {
+  static AppointmentUpdatingMode create(HeightUpdateFrom from) {
+    switch(from){
+      case HeightUpdateFrom.top:
+        return AppointmentUpdatingMode.changeTop;
+      case HeightUpdateFrom.bottom:
+        return AppointmentUpdatingMode.changeHeight;
+    }
+  }
+}