+++ lib/src/ui/interval_containers/appointment_container/drag_indicator_widget.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 
 class DragIndicatorWidget extends StatelessWidget {
   const DragIndicatorWidget.top({
-    required this.child,
     required this.onLongPressDown,
     required this.onLongPressStart,
     required this.onLongPressEnd,
@@ -15,7 +14,6 @@ class DragIndicatorWidget extends StatelessWidget {
         super(key: key);
 
   const DragIndicatorWidget.bottom({
-    required this.child,
     required this.onLongPressDown,
     required this.onLongPressStart,
     required this.onLongPressEnd,
@@ -27,12 +25,35 @@ class DragIndicatorWidget extends StatelessWidget {
   })  : mode = _Mode.bottom,
         super(key: key);
 
+  const DragIndicatorWidget.overlayTop({
+    this.dragIndicatorBorderColor,
+    this.dragIndicatorBorderWidth,
+    this.dragIndicatorColor,
+    Key? key,
+  })  : mode = _Mode.overlayTop,
+        onLongPressDown = null,
+        onLongPressStart = null,
+        onLongPressMoveUpdate = null,
+        onLongPressEnd = null,
+        super(key: key);
+
+  const DragIndicatorWidget.overlayBottom({
+    this.dragIndicatorBorderColor,
+    this.dragIndicatorBorderWidth,
+    this.dragIndicatorColor,
+    Key? key,
+  })  : mode = _Mode.overlayBottom,
+        onLongPressDown = null,
+        onLongPressStart = null,
+        onLongPressMoveUpdate = null,
+        onLongPressEnd = null,
+        super(key: key);
+
   final _Mode mode;
-  final Widget child;
-  final GestureLongPressCallback onLongPressDown;
-  final GestureLongPressStartCallback onLongPressStart;
-  final GestureLongPressEndCallback onLongPressEnd;
-  final GestureLongPressMoveUpdateCallback onLongPressMoveUpdate;
+  final GestureLongPressCallback? onLongPressDown;
+  final GestureLongPressStartCallback? onLongPressStart;
+  final GestureLongPressEndCallback? onLongPressEnd;
+  final GestureLongPressMoveUpdateCallback? onLongPressMoveUpdate;
 
   ///The color to be applied to the default drag indicator widget.
   final Color? dragIndicatorColor;
@@ -45,52 +66,52 @@ class DragIndicatorWidget extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Stack(
-      clipBehavior: Clip.none,
-      children: [
-        child,
-        Positioned(
-          top: mode == _Mode.top ? -2 : null,
-          right: 0,
-          left: 0,
-          bottom: mode == _Mode.bottom ? -2 : null,
-          child: GestureDetector(
-            onLongPress: onLongPressDown,
-            onLongPressStart: onLongPressStart,
-            onLongPressEnd: onLongPressEnd,
-            onLongPressMoveUpdate: onLongPressMoveUpdate,
-            behavior: HitTestBehavior.opaque,
-            child: Container(
-              alignment: mode == _Mode.bottom
-                  ? Alignment.centerRight
-                  : Alignment.centerLeft,
-              padding: const EdgeInsets.symmetric(horizontal: 10),
-              child: Container(
-                clipBehavior: Clip.hardEdge,
-                width: 10,
-                height: 10,
-                decoration: BoxDecoration(
-                    shape: BoxShape.circle,
-                    color: dragIndicatorColor ?? Colors.white,
-                    border: Border.all(
-                      color: dragIndicatorBorderColor ??
-                          Theme.of(context).primaryColor,
-                      width: dragIndicatorBorderWidth ?? 3,
-                    ),
-                    boxShadow: const [
-                      BoxShadow(
-                        offset: Offset(1, 1),
-                        blurRadius: 1,
-                        color: Colors.black45,
-                      )
-                    ]),
-              ),
+    final Widget indicatorWidget = _buildContainer(context);
+    return Positioned(
+      top: [_Mode.top, _Mode.overlayTop].contains(mode) ? -2 : null,
+      right: 0,
+      left: 0,
+      bottom: [_Mode.bottom, _Mode.overlayBottom].contains(mode) ? -2 : null,
+      child: [_Mode.overlayBottom, _Mode.overlayTop].contains(mode)
+          ? indicatorWidget
+          : GestureDetector(
+              onLongPress: onLongPressDown,
+              onLongPressStart: onLongPressStart,
+              onLongPressEnd: onLongPressEnd,
+              onLongPressMoveUpdate: onLongPressMoveUpdate,
+              behavior: HitTestBehavior.opaque,
+              child: indicatorWidget,
             ),
-          ),
-        ),
-      ],
+    );
+  }
+
+  Widget _buildContainer(BuildContext context) {
+    return Container(
+      alignment: [_Mode.bottom, _Mode.overlayBottom].contains(mode)
+          ? Alignment.centerRight
+          : Alignment.centerLeft,
+      padding: const EdgeInsets.symmetric(horizontal: 10),
+      child: Container(
+        clipBehavior: Clip.hardEdge,
+        width: 10,
+        height: 10,
+        decoration: BoxDecoration(
+            shape: BoxShape.circle,
+            color: dragIndicatorColor ?? Colors.white,
+            border: Border.all(
+              color: dragIndicatorBorderColor ?? Theme.of(context).primaryColor,
+              width: dragIndicatorBorderWidth ?? 3,
+            ),
+            boxShadow: const [
+              BoxShadow(
+                offset: Offset(1, 1),
+                blurRadius: 1,
+                color: Colors.black45,
+              )
+            ]),
+      ),
     );
   }
 }
 
-enum _Mode { top, bottom }
+enum _Mode { top, bottom, overlayTop, overlayBottom }